Fixes case where package has multi versions of same dep
It's possible for a root package to have multiple versions of the same
dependency. For example Can has a devDependency on jquery and a
peerDependency on jquery. Unless both are the same semver range loading
the config will throw when trying to convert the name. This fixes it by
not making the assumption that there is only 1 version of a given
dependency and instead does a semver check to see. Fixes #14