Support browser's falsey option
`browser` can be used to ignore a module. There seemed to be partial
support for this functionality in the existing code but it didn't
actually work because normalize always wants to find a package, but
there won't be one for something like `@empty`. The fix is to make
`refPkg.browser` a sort of fall-through mapping for modules when a
package isn't found.