Makes contextual maps work
Contextual mappings were not working previously because when we
converting map names to npm-like names we were assuming the value was
also a string that could be converted, as is commonly the case. In the
case of contextual maps the value will be an object. Fixes #50

Also added a test that demonstrates the ability to feed a dependency one
of its own dependencies while mocking that dependency to the rest of the
work.