configDependencies can export config with systemConfig
This adds the ability for configDependencies to export configuration
using a special export called `systemConfig`. For example:

```js
exports.systemConfig = {
	map: {
		one: "two"
	}
};
```

This will run this object through `loader.config` after the extension
has been applied, so names are properly converted.

Not documenting for now as I think there are better solutions to this
problem long term.

Fixes #52