Fix npm traversal after missing
When npm3 misses (looks for a package.json and doesn't find it) the
traversal is slightly different from npm2. In npm2 you would traverse UP
always. But with npm3 you are most likely starting from a less nested
position. So to traverse you start from the most nested position and
*then* traverse up.