Leave pkgs in context untouched
When crawling an npm3 file structure, don't overwrite packages in the
context. Previously whenever we found a wrong version we would keep a
reference to that pkg object and then overwrite the object when we found
the correct version. This meant that our context contained duplicate
objects, and some versions were incorrect.

Closes #92 and #72