Load modules that use folder/index.js convention
This adds support for modules that use the `folder/index.js` convention
which is common in Node code. The strategy to make this work is to add a
fetch hook that, when loading fails, retries with `folder/index.js`. If
this is successful it will update the `load.address` to be correct, and
creates a `npmParentMap` entry that looks like:

```
{
  "folder": "folder/index"
}
```

This is used in normalize in order to correctly normalize relative
modules. So if there is a module:

```
./utils
```

It will be correctly normalized to `folder/utils`.

This fixes #3