Refactor so that saving package.json happens outside npm.js
This refactor moves the logic that creates the source for the
`package.json!npm` module into a new module, `npm-load`. This module
contains the function `saveLoadIfNeeded` which will recreate the source
on-demand during the build. This will be used for 2 scenarios
(currently):

1. If we find a module that uses the `folder/index.js` convention, this
information is saved into a side table that needs to be recreated in
production.

2. So that we can progressively load package.jsons.