Progressively load package.jsons
This implements progressively loading package.json files.

This is a big change. There are two main things to be aware of:

1. Previously we crawled all package.jsons within the project, going as
deep as possible. Now we only crawl **first level dependencies**, that
is dependencies (including dev/peer) of the top-level project.

The rest of the packages are progressively loaded as needed in the
`normalize` hook.

2. Conversion of config had to change because of this. Previously since
all package.jsons were loaded up-front we would then convert all config
over to npm names after all package.jsons were loaded. However this
won't work now, some packages are not available to be converted.

So when we try to convert config we detect that a package is not
available and push it into a waiting queue. Every time a new
package.json is loaded we check the queue to see if any config was
waiting on it, and if so we rerun that configuration.

Closes #41