Reuse existing pkgInfo if it exists
During the build we will be loading separate bundles that have their own
set of package dependencies. Because for each bundle we will be creating
a separate `package.json!npm` source, none of those will encompass ALL
packages, which is what we need for the build.

Instead each time we load the graph for different bundles we'll reuse
the same pkgInfo (this contains the minimal data needed for each
		package) so that we build a full `package.json!npm` module that
contains all of the source needed.

This is for https://github.com/stealjs/steal-tools/issues/365