Progressively load package.jsons
This implements progressively loading package.json files.

This is a big change. There are two main things to be aware of:

1. Previously we crawled all package.jsons within the project, going as
deep as possible. Now we only crawl **first level dependencies**, that
is dependencies (including dev/peer) of the top-level project.

The rest of the packages are progressively loaded as needed in the
`normalize` hook.

2. Conversion of config had to change because of this. Previously since
all package.jsons were loaded up-front we would then convert all config
over to npm names after all package.jsons were loaded. However this
won't work now, some packages are not available to be converted.

So when we try to convert config we detect that a package is not
available and push it into a waiting queue. Every time a new
package.json is loaded we check the queue to see if any config was
waiting on it, and if so we rerun that configuration.

Closes #41

Load modules that use folder/index.js convention

This adds support for modules that use the `folder/index.js` convention
which is common in Node code. The strategy to make this work is to add a
fetch hook that, when loading fails, retries with `folder/index.js`. If
this is successful it will update the `load.address` to be correct, and
creates a `npmParentMap` entry that looks like:

```
{
	  "folder": "folder/index"
}
```

This is used in normalize in order to correctly normalize relative
modules. So if there is a module:

```
./utils
```

It will be correctly normalized to `folder/utils`.

This fixes #3