Always save version range configuration after loading a package.json
We have a map that uses a versionRange as the key and a package.json as
the value. In a scenario where we have to load more than one
package.json (because one in the wrong place) we need to make sure this
versions map is updated so that config is correctly applied.

Fixes #694