Refactor package.json fetching
This refactors the package.json fetching to be a bit more readable. This
creates a `FetchTask` object. This object represents one *attempt* at
fetching the package.json. It may fail because:

1) A 404.
2) We fetched a package.json that is not a matching version.

If either of these things occur you can call `task.next()` which will
return a new `pkg` object that represents the next location to fetch
from (where that is depends on if using npm3, but eventually we traverse
		up the tree).

This makes the core fetching logic easier to read. Attempt a fetch, if
it fails get a new pkg, and recurse (creating a new `FetchTask` each
		time).