Correctly normalize ..
This updates utils.moduleName.parseFromPackage to correctly parse
a name ".." that is coming from a parent package. We already had logic
for ../ and ./, so just needed to adjust it to work with ..

Additionally moved this logic into utils.moduleName.parseFromPackage
where it belongs, rather than directly in the normalize hook.

For https://github.com/stealjs/steal/issues/770