Restore system.main after applying config
A couple of places we need to run a package.json' `system` property
through `loader.config()` in order to configure the loader. Whenever we
do this we remove the `system.main` property so that `System.main` is
not set (this will be the app's "main"). In one place we forgot to
restore this property afterwards, as it's needed to properly normalize
in production.

Fixes #168