Config in parent package should take preference to child packages
This fixes #185

Changes the way package config is applied, now in reverse order, so that
the root package's config is applied last, making it so that it
overrides any child config.

This doesn't fix the issue of progressively loaded configuration, that
problem is probably unfixable as long as we allow child packages to
override global configuration.

Closes #185