Don't override package.json config
Now so that we can support progressive loading of config that is applied
via loader.config() before the module has been normalized we are placing
that config in a waiting queue. We don't want that config to apply the
root package.json config, because the package.json config is what should
be written out into the main bundle.