Retain steal configuration from previous bundles
When steal-tools runs its build it loads each bundle separately a new
loader like so:

- a
- b
- c

If `b` has some configuration needed for its production bundle and `c`
runs through without that configuration, that configuration will be
omited.

This fixes it so that each time we load a new bundle we use the previous
bundles configuration as a starting point.

Closes #216