Remove live-reload code
In Steal 0.16 we needed special live-reload behavior because when
steal-npm was imported it would fetch all of the package.jsons. Now that
we only fetch package.jsons as they are needed there isn't the need to
flush out the old state, as the state remains valid.

This is the cause of https://github.com/donejs/donejs/issues/790